Got everything working with autoupdates

+++ frontend/src/app/board/board.component.ts
@@ -3,7 +3,6 @@ import { CanvasWhiteboardUpdate, CanvasWhiteboardModule, CanvasWhiteboardShapeOp
 import { CanvasWhiteboardComponent } from 'ng2-canvas-whiteboard';
 import { BoardService } from '../board.service';
 import { Board } from '../board';
-import { Router, NavigationEnd } from '@angular/router';
 
 @Component({
   selector: 'app-board',
@@ -13,41 +12,45 @@ import { Router, NavigationEnd } from '@angular/router';
 })
 export class BoardComponent implements OnInit, OnDestroy {
 
-  constructor(public boardService: BoardService) {//, private router: Router) {
-    // subscribe to the router events - storing the subscription so
-    // we can unsubscribe later.    
-    // this.navigationSubscription = this.router.events.subscribe((e: any) => {
-    //   // If it is a NavigationEnd event re-initalise the component
-    //   if (e instanceof NavigationEnd) {
-    //   }
-    // });
+  constructor(public boardService: BoardService) {
   }
 
   @ViewChild('canvasWhiteboard') canvasWhiteboard: CanvasWhiteboardComponent;
   currentBoard: Board = new Board;
-  drawnUUIDs = [];
+  drawnUpdates = [];
   sendUpdates: CanvasWhiteboardUpdate[] = [];
-  // databaseEmptied: boolean = false;
-  // navigationSubscription;
+  databaseEmptied: boolean = false;
+  drawing: boolean = false;
+  deleting: boolean = false;
+  databaseUpdating: boolean = false;
 
   sendBatchUpdate(updates: CanvasWhiteboardUpdate[]) {
-    // console.log(updates)
-    let board = new Board()
-    board.canvasWhiteboardUpdates = updates
-    this.boardService.save(board).subscribe()
-    console.log(this.canvasWhiteboard.getDrawingHistory())
+    if (!this.databaseUpdating) {
+      this.drawing = true;
+        let board = new Board();
+        board.canvasWhiteboardUpdates = updates;
+      this.boardService.save(board).subscribe(data => {
+        var lastUpdate = updates[updates.length - 1];
+        if (lastUpdate.type == 2) {
+          this.drawing = false;
+        }
+      })
+      console.log(this.canvasWhiteboard.getDrawingHistory());
+    }
   }
 
   onCanvasClear() {
-    this.printDebugInfo()
-    this.boardService.delete(this.currentBoard.id).subscribe(data => {
-      console.log("database was emptied.")
-    })
-    this.sendUpdates = []
-    this.drawnUUIDs = []
-    this.currentBoard = new Board()
-    console.log("The canvas was cleared");
-    // this.router.navigateByUrl('/board')
+    if (!this.databaseUpdating) {
+      this.deleting = true;
+      this.boardService.delete(this.currentBoard.id).subscribe(data => {
+        this.sendUpdates = [];
+        this.drawnUpdates = [];
+        this.currentBoard = new Board();
+        this.deleting = false;
+        console.log("Database was emptied.");
+      })
+      console.log("The canvas was cleared");
+    }
   }
 
   onCanvasUndo(updateUUID: string) {
@@ -55,99 +58,96 @@ export class BoardComponent implements OnInit, OnDestroy {
   }
 
   onCanvasRedo(updateUUID: string) {
-
     console.log(`REDO with uuid: ${updateUUID}`);
   }
 
   drawDatabaseUpdates() {
-    // this.canvasWhiteboard.drawUpdates([new CanvasWhiteboardUpdate(0.35661764705882354, 0.383399209486166, 0, "f5e3ea98-362b-de26-eff3-f78938ed3a0e", "FreeHandShape")]);
-    // this.canvasWhiteboard.drawUpdates([new CanvasWhiteboardUpdate(0.35661764705882354, 0.383399209486166, 2, "f5e3ea98-362b-de26-eff3-f78938ed3a0e", "FreeHandShape")]);
-    // first check if mouse is being used
-    // if (!this.canvasWhiteboard.drawingEnabled) {
-    console.log("refreshing from db!");
-    this.boardService.find(this.currentBoard.id).subscribe(data => {
-      if (data != null) {
-        console.log("refreshing from db! [data != null]");
-        // now check if updates dont contain undefined
-        // this.currentBoard = data
-        if (data.canvasWhiteboardUpdates != null) {
-          if (data.canvasWhiteboardUpdates.length > 0) {
-            var allCorrect: boolean = true;
-            var endIndexes: number[] = [0];
-            var cleanUpdates: CanvasWhiteboardUpdate[] = [];
-            data.canvasWhiteboardUpdates.forEach((element, index) => {
-              if (element == undefined || element == null) {
-                allCorrect = false;
-              } else if (element.type == 2 || element.type == 1) {
-                endIndexes.push(index);
-                cleanUpdates.push(this.cleanUpUpdate(element))
-              } else{
-                cleanUpdates.push(this.cleanUpUpdate(element))
-              }
-            });
-            if (allCorrect) {
-              console.log("refreshing from db! [allCorrect]");
-              // console.log(this.canvasWhiteboard.getDrawingHistory())
-              var sendUpdates: CanvasWhiteboardUpdate[] = []
-              var newUUIDs = []
-              var finishedUpdatesSlice: CanvasWhiteboardUpdate[] = cleanUpdates.slice(0, endIndexes[endIndexes.length - 1])
-              // add only if not drawn yet.
-              finishedUpdatesSlice.forEach(childUpdate => {
-                if (!this.canvasWhiteboard.getDrawingHistory().some((parentUpdate) => (
-                  childUpdate.UUID == parentUpdate.UUID &&
-                  childUpdate.x == parentUpdate.x &&
-                  childUpdate.y == parentUpdate.y &&
-                  childUpdate.selectedShape == parentUpdate.selectedShape &&
-                  childUpdate.type == parentUpdate.type
-                ))) {
-                  sendUpdates.push(childUpdate)
+    // first check if mouse is being used or a delete action is happening
+    if (!this.drawing && !this.deleting ){
+      //&& !this.canvasWhiteboard.drawingEnabled) {
+      this.drawing = false; 
+      this.deleting = false;
+      this.databaseUpdating = true;
+      console.log("refreshing from db!");
+      this.boardService.find(this.currentBoard.id).subscribe(data => {
+        if (data != null) {
+          console.log("refreshing from db! [data != null]");
+          // now check if updates dont contain undefined
+          // this.currentBoard = data
+          if (data.canvasWhiteboardUpdates != null) {
+            this.databaseEmptied = false;
+            if (data.canvasWhiteboardUpdates.length > 0) {
+              var allCorrect: boolean = true;
+              var endIndexes: number[] = [0];
+              var cleanUpdates: CanvasWhiteboardUpdate[] = [];
+              data.canvasWhiteboardUpdates.forEach((element, index) => {
+                if (element == undefined || element == null) {
+                  allCorrect = false;
+                } else if (element.type == 2) {
+                  endIndexes.push(index+1);
+                  cleanUpdates.push(this.cleanUpUpdate(element));
+                } else {
+                  cleanUpdates.push(this.cleanUpUpdate(element));
                 }
-              })
+              });
+              if (allCorrect) {
+                console.log("refreshing from db! [allCorrect]");
+                // console.log(this.canvasWhiteboard.getDrawingHistory())
+                var sendUpdates: CanvasWhiteboardUpdate[] = [];
+                var finishedUpdatesSlice: CanvasWhiteboardUpdate[] = cleanUpdates.slice(0, endIndexes[endIndexes.length-1]);
+                
+                sendUpdates = finishedUpdatesSlice;
+                
+                
+                // // add only if not drawn yet.
 
-              if (sendUpdates.length > 0) {
-                console.log("new!")
-                this.sendUpdates = sendUpdates
-              } else {
-                this.sendUpdates = []
-                console.log("nothing new!")
+                // finishedUpdatesSlice.forEach(childUpdate => {
+                //   if (!this.drawnUpdates.some((parentUpdate) => (
+                //     childUpdate.UUID == parentUpdate.UUID &&
+                //     childUpdate.x == parentUpdate.x &&
+                //     childUpdate.y == parentUpdate.y &&
+                //     childUpdate.selectedShape == parentUpdate.selectedShape &&
+                //     childUpdate.type == parentUpdate.type
+                //   ))) {
+                //     sendUpdates.push(childUpdate);
+                //   }
+                // })
+
+                if (sendUpdates.length > 0) {
+                  // console.log("new update!");
+                  this.sendUpdates = sendUpdates;
+                  this.drawPerUUID(this.sendUpdates)
+                  this.drawnUpdates.push(...this.sendUpdates)
+                  this.sendUpdates = []
+                } else {
+                  // this.sendUpdates = [];
+                  // console.log("nothing new!");
+                }
               }
+            } else {
+              this.databaseEmptied = true;
+              this.canvasWhiteboard.clearCanvas();
             }
           }
-          // } else {
-          //   this.databaseEmptied = true;
-          // }
-
         }
-      }
-    })
+        this.databaseUpdating = false;
+      })
+    }
   }
 
   ngOnInit(): void {
-    // history not registered with foreign drawUpdates?... (BUG)
     setInterval(() => {
-      this.printDebugInfo()
       // console.log("interval")
-      // console.log(console.log(this.canvasWhiteboard.getDrawingHistory()))
+      // this.printDebugInfo()
       this.drawDatabaseUpdates();
-      // this.canvasWhiteboard.clearCanvas()
-      this.drawPerUUID(this.sendUpdates)
-      this.sendUpdates = []
-      // if (this.databaseEmptied) {
-      //   this.router.navigateByUrl('/board')
-      // }
-    }, 500);
+    }, 1000);
   }
 
   ngOnDestroy() {
-    // avoid memory leaks here by cleaning up after ourselves. If we  
-    // don't then we will continue to run our initialiseInvites()   
-    // method on every navigationEnd event.
-    // if (this.navigationSubscription) {
-    //   this.navigationSubscription.unsubscribe();
-    // }
   }
 
   cleanUpUpdate(update: CanvasWhiteboardUpdate) {
+    // cleans up updates (reinitializes them with only important values)
     // check if shapeoptions is defined.
     var newShapeOptions = new CanvasWhiteboardShapeOptions()
     if (update.selectedShapeOptions != null) {
@@ -162,21 +162,38 @@ export class BoardComponent implements OnInit, OnDestroy {
   }
 
   drawPerUUID(updates: CanvasWhiteboardUpdate[]) {
-    var updateSet: CanvasWhiteboardUpdate[] = []; 
+    // this assumes the updates are in correct order!
+    var updateSet: CanvasWhiteboardUpdate[] = [];
     updates.forEach(element => {
-      updateSet.push(element)  
+      updateSet.push(element)
       if (element.type == 2) {
-          this.canvasWhiteboard.drawUpdates(updateSet)
-          updateSet = []
-        }
+        console.log("finished!")
+        this.canvasWhiteboard.drawUpdates(updateSet);
+        updateSet = [];
+      }
     })
   }
 
   printDebugInfo() {
-    console.log(
-      "===DrawingEnabled===;\n"+JSON.stringify(this.canvasWhiteboard.getDrawingEnabled())+
-      "\n===ShouldDraw===;\n"+JSON.stringify(this.canvasWhiteboard.getShouldDraw())+
-      "\n===DrawingHistory===;\n"+JSON.stringify(this.canvasWhiteboard.getDrawingHistory())
-    )
+    // console.log(
+    //   "===DrawingEnabled===;\n" + JSON.stringify(this.canvasWhiteboard.getDrawingEnabled()) +
+    //   "\n===ShouldDraw===;\n" + JSON.stringify(this.canvasWhiteboard.getShouldDraw()) +
+    //   "\n===DrawingHistory===;\n" + JSON.stringify(this.canvasWhiteboard.getDrawingHistory())
+    // );
+
+    console.log(this.drawing)
+    console.log(this.deleting)
+    console.log(this.canvasWhiteboard.drawingEnabled)
+    console.log(this.databaseUpdating)
+    console.log(this.databaseEmptied)
+  }
+
+  testButton() {
+    this.canvasWhiteboard.drawUpdates([
+      new CanvasWhiteboardUpdate(0.35661764705882354, 0.383399209486166, 0, "f5e3ea98-362b-de26-eff3-f78938ed3a0e", "FreeHandShape"),
+      new CanvasWhiteboardUpdate(0.35661764705882354, 0.383399209486166, 2, "f5e3ea98-362b-de26-eff3-f78938ed3a0e", "FreeHandShape"),
+      new CanvasWhiteboardUpdate(0.45661764705882354, 0.403399209486166, 0, "f6e3ea98-362b-de26-eff3-f78938ed3a0f", "FreeHandShape"),
+      new CanvasWhiteboardUpdate(0.45661764705882354, 0.403399209486166, 2, "f6e3ea98-362b-de26-eff3-f78938ed3a0f", "FreeHandShape")
+    ]);
   }
 }
